+++ core/src/test/java/com/bc/fiduceo/db/DbMaintenanceToolIntegrationTest.java
@@ -65,7 +65,7 @@ public class DbMaintenanceToolIntegrationTest {
             psE.flush();
 
             assertEquals("", out.toString());
-            assertEquals("db-maintenance-tool version 1.4.4-SNAPSHOT" + ls +
+            assertEquals("db-maintenance-tool version 1.4.4" + ls +
                     ls +
                     "usage: db-maintenance-tool <options>" + ls +
                     "Valid options are:" + ls +