+++ core/src/test/java/com/bc/fiduceo/db/DbMaintenanceToolTest.java
@@ -52,7 +52,7 @@ public class DbMaintenanceToolTest {
 
         tool.printUsageTo(stream);
 
-        assertEquals("db-maintenance-tool version 1.4.4-SNAPSHOT" + ls +
+        assertEquals("db-maintenance-tool version 1.4.4" + ls +
                 ls +
                 "usage: db-maintenance-tool <options>" + ls +
                 "Valid options are:" + ls +