+++ matchup-tool/src/test/java/com/bc/fiduceo/matchup/MatchupToolIntegrationTest.java
@@ -34,7 +34,7 @@ import static org.junit.Assert.assertEquals;
 public class MatchupToolIntegrationTest {
 
     private final String ls = System.lineSeparator();
-    private final String expectedPrintUsage = "matchup-tool version 1.4.4-SNAPSHOT" + ls +
+    private final String expectedPrintUsage = "matchup-tool version 1.4.4" + ls +
             ls +
             "usage: matchup-tool <options>" + ls +
             "Valid options are:" + ls +