+++ matchup-tool/src/test/java/com/bc/fiduceo/matchup/MatchupToolTest.java
@@ -84,7 +84,7 @@ public class MatchupToolTest {
 
         matchupTool.printUsageTo(outputStream);
 
-        assertEquals("matchup-tool version 1.4.4-SNAPSHOT" + ls +
+        assertEquals("matchup-tool version 1.4.4" + ls +
                 ls +
                 "usage: matchup-tool <options>" + ls +
                 "Valid options are:" + ls +